+++ css-library/neo/changelog.md
@@ -1,5 +1,31 @@
 # Changelog
 
+### 3.72.0
+
+### Breaking Changes - Needs Immediate Attention
+
+- N/A
+
+### Additional Changes - The following items have been updated. Please make these updates at your earliest convenience.
+
+- N/A
+
+### New Features - New features that have recently been added
+
+- Added "loop-single", "loop-multiple", "variable-assign", "sub-accounts-select", "sub-accounts-switch", "customer-journey-retrieve-data", "customer-journey-store-data", "label-go", "broadcast-play", "broadcast-stop", "callback-offer", "queue-removed", icons (NEO-1481)
+
+### Bugs Report - List of bugs that have been updated
+
+- N/A
+
+### Sunsetting - Removing Full Support
+
+- N/A
+
+### Work In Progress - Future Updates
+
+- N/A
+
 ### 3.71.0
 
 ### Breaking Changes - Needs Immediate Attention